ДЗ 5. Не справился. МФУ, реализовано копирование через отдельный метод синхронизованный по скану и печати (как по моему колхоз) скорее всего можно было ка-то обойтись без доп.метод, но не понял как (но пример рабочий). С гонками не справился, так и непонял что и как использовать...
