+++ Race/src/Car.java
@@ -1,0 +1,39 @@
+import java.util.concurrent.Semaphore;
+
+public class Car implements Runnable {
+    private static int CARS_COUNT;
+    //Semaphore smp = new Semaphore(0);
+    static {
+        CARS_COUNT = 0;
+    }
+    private Race race;
+    private int speed;
+    private String name;
+    public String getName() {
+        return name;
+    }
+    public int getSpeed() {
+        return speed;
+    }
+    public Car(Race race, int speed) {
+        this.race = race;
+        this.speed = speed;
+        CARS_COUNT++;
+        this.name = "Участник #" + CARS_COUNT;
+    }
+    @Override
+    public void run() {
+        try {
+            System.out.println(this.name + " готовится");
+            Thread.sleep(500 + (int)(Math.random() * 800));
+            System.out.println(this.name + " готов");
+            //smp.acquire(); //блокируется совсем потом не отпускает потоки
+            } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        for (int i = 0; i < race.getStages().size(); i++) {
+            race.getStages().get(i).go(this);
+        }
+    }
+}
