+++ Race/src/MainClass.java
@@ -1,0 +1,25 @@
+//import java.util.concurrent.CyclicBarrier;
+
+public class MainClass {
+
+    public static final int CARS_COUNT = 4;
+
+    public static void main(String[] args) {
+        System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Подготовка!!!");
+        Race race = new Race(new Road(60), new Tunnel(), new Road(40));
+        Car[] cars = new Car[CARS_COUNT];
+
+        for (int i = 0; i < cars.length; i++) {
+            cars[i] = new Car(race, 20 + (int) (Math.random() * 10));
+        }
+        for (int i = 0; i < cars.length; i++) {
+            new Thread(cars[i]).start();
+        }
+
+        System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка началась!!!"); //должно появляться после того как все гонки готовы
+        // перед стартом и происходит старт
+
+        System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка закончилась!!!"); //должно появиться после того как все
+        // гонки закончат последний этап
+    }
+}
