+++ Race/src/Stage.java
@@ -1,0 +1,8 @@
+public abstract class Stage {
+    protected int length;
+    protected String description;
+    public String getDescription() {
+        return description;
+    }
+    public abstract void go(Car c);
+}
